better range type conversation cleanup introduced TypeRef and VarRef(Simple) in EVL and PIR

+++ src/evl/traverser/range/CaseRangeUpdater.java
@@ -30,7 +30,7 @@ public class CaseRangeUpdater extends NullTraverser<Void, Map<Variable, Range>> 
 
   private BigInteger evalToInt(Expression value) {
     assert (value instanceof Number);
-    return BigInteger.valueOf(((Number) value).getValue());
+    return ((Number) value).getValue();
   }
 
   private void setHigh(BigInteger num) {
